+++ R/fill_dir.R
@@ -43,12 +43,17 @@
 #'  A \code{list} of a single model's script-writing controls or a
 #'  \code{list} of \code{list}s, each of which is a single model's 
 #'  script-writing controls. \cr 
-#'  Presently, each model's script writing controls
-#'  should include three elements: \code{name} (a \code{character} value of 
-#'  the model name), \code{covariates} (a \code{logical} indicator of if the 
-#'  model needs covariates), and \code{lag} (an \code{integer}-conformable 
-#'  value of the lag to use with the covariates or \code{NA} if 
-#'  \code{covariates = FALSE}). \cr 
+#'  Presently, each model's script writing controls should include four 
+#'  elements: 
+#'  \itemize{
+#'   \item \code{name}: a \code{character} value of the model name.
+#'   \item \code{data_sets}: a \code{character} vector of the data set names
+#'    that the model is applied to. 
+#'   \item \code{covariatesTF}: a \code{logical} indicator of if the 
+#'    model needs covariates.
+#'   \item \code{lag}: an \code{integer}-conformable value of the lag to use 
+#'    with the covariates or \code{NA} if \code{covariatesTF = FALSE}.
+#'  } 
 #'  If only a single model is added, the name of 
 #'  the model from the element \code{name} will be used to name the model's
 #'  \code{list} in the larger \code{list}. If multiple models are added, each
@@ -72,6 +77,13 @@
 #'
 #' @param quiet \code{logical} indicator if progress messages should be
 #'  quieted.
+#'
+#' @param only_if_missing \code{logical} indicator if the content should
+#'  be filled only if it is missing. Only used in \code{fill_raw}, where
+#'  missingness is determined by \code{\link{verify_raw_data}}. Default
+#'  is \code{FALSE}, which allows for control via \code{link{files_control}}'s
+#'  \code{overwrite} argument. Only accessed and changed to \code{FALSE} 
+#'  within \code{\link{portalcast}}.  
 #'
 #' @param verbose \code{logical} indicator of whether or not to print out
 #'  all of the information or not (and thus just the tidy messages). 
@@ -252,6 +264,7 @@ fill_casts <- function(main = ".", control_files = files_control(),
 #'
 fill_raw <- function(main = ".", 
                      downloads = zenodo_downloads(c("1215988", "833438")), 
+                     only_if_missing = FALSE,
                      control_files = files_control(), quiet = FALSE, 
                      arg_checks = TRUE){
   check_args(arg_checks = arg_checks)
@@ -259,6 +272,14 @@ fill_raw <- function(main = ".",
   if(list_depth(downloads) == 1){
     downloads <- list(downloads)
   }
+
+  raw_data <- control_files$raw_data
+  raw_data_present <- verify_raw_data(main = main, raw_data = raw_data, 
+                                      arg_checks = arg_checks)
+  if(raw_data_present & only_if_missing){
+    return(invisible(NULL))
+  }
+
   messageq(" -Downloading raw files", quiet)
   ndl <- length(downloads)
   dl_vers <- rep(NA, ndl)
@@ -274,3 +295,4 @@ fill_raw <- function(main = ".",
                           quiet = quiet, arg_checks = arg_checks)
   invisible(NULL)
 }
+